Sanitize exceptions
