+++ sdk/core/Azure.Core/src/Azure.Core.csproj
@@ -25,9 +25,11 @@
     <Compile Include="Shared\Argument.cs" />
     <Compile Include="Shared\EventSourceEventFormatting.cs" />
     <Compile Include="Shared\HashCodeBuilder.cs" />
+    <Compile Include="Shared\ClientDiagnostics.cs" />
+    <Compile Include="Shared\HttpMessageSanitizer.cs" />
     <Compile Include="Shared\NullableAttributes.cs" />
     <Compile Include="Shared\ContentTypeUtilities.cs" />
-    <Compile Include="Shared\ClientDiagnostics.cs" />
+    <Compile Include="Shared\DiagnosticsScopeFactory.cs" />
     <Compile Include="Shared\DiagnosticScope.cs" />
   </ItemGroup>
 </Project>
