+++ sdk/core/Azure.Core/tests/Azure.Core.Tests.csproj
@@ -27,7 +27,6 @@
     <Compile Include="..\src\Shared\ForwardsClientCallsAttribute.cs" />
     <Compile Include="..\src\Shared\RetriableStream.cs" />
     <Compile Include="..\src\Shared\NoBodyResponse{T}.cs" />
-    <Compile Include="..\src\Shared\ResponseExceptionExtensions.cs" />
     <Compile Include="..\src\Shared\OperationHelpers.cs" />
     <Compile Include="..\src\Shared\AzureResourceProviderNamespaceAttribute.cs" />
   </ItemGroup>
