+++ sdk/identity/Azure.Identity/src/ManagedIdentityClient.cs
@@ -75,7 +75,7 @@ namespace Azure.Identity
             {
                 _msiType = MsiType.Unavailable;
 
-                ValueTask<string> messageTask = ResponseExceptionExtensions.CreateRequestFailedMessageAsync(IdentityUnavailableError, response, null, false);
+                ValueTask<string> messageTask = _pipeline.Diagnostics.CreateRequestFailedMessageAsync(IdentityUnavailableError, response, null, false);
 
                 // TODO: this should use TaskExtensions EnsureCompleted from Azure.Core shared source when it gets move into shared source.
                 Debug.Assert(messageTask.IsCompleted);
@@ -85,7 +85,7 @@ namespace Azure.Identity
                 return new ExtendedAccessToken(new CredentialUnavailableException(message));
             }
 
-            throw response.CreateRequestFailedException();
+            throw _pipeline.Diagnostics.CreateRequestFailedException(response);
         }
 
         public async virtual Task<ExtendedAccessToken> AuthenticateAsync(string[] scopes, CancellationToken cancellationToken)
@@ -113,12 +113,12 @@ namespace Azure.Identity
             {
                 _msiType = MsiType.Unavailable;
 
-                string message = await ResponseExceptionExtensions.CreateRequestFailedMessageAsync(IdentityUnavailableError, response, null, true).ConfigureAwait(false);
+                string message = await _pipeline.Diagnostics.CreateRequestFailedMessageAsync(IdentityUnavailableError, response, null, true).ConfigureAwait(false);
 
                 return new ExtendedAccessToken(new CredentialUnavailableException(message));
             }
 
-            throw await response.CreateRequestFailedExceptionAsync().ConfigureAwait(false);
+            throw await _pipeline.Diagnostics.CreateRequestFailedExceptionAsync(response).ConfigureAwait(false);
         }
 
         protected virtual MsiType GetMsiType(CancellationToken cancellationToken)
