+++ sdk/keyvault/Azure.Security.KeyVault.Secrets/src/RecoverDeletedSecretOperation.cs
@@ -105,7 +105,7 @@ namespace Azure.Security.KeyVault.Secrets
         public override ValueTask<Response<SecretProperties>> WaitForCompletionAsync(TimeSpan pollingInterval, CancellationToken cancellationToken) =>
             this.DefaultWaitForCompletionAsync(pollingInterval, cancellationToken);
 
-        private static async ValueTask<bool> CheckCompletedAsync(Response response)
+        private async ValueTask<bool> CheckCompletedAsync(Response response)
         {
             switch (response.Status)
             {
@@ -117,10 +117,10 @@ namespace Azure.Security.KeyVault.Secrets
                     return false;
 
                 default:
-                    throw await response.CreateRequestFailedExceptionAsync().ConfigureAwait(false);
+                    throw await _pipeline.Diagnostics.CreateRequestFailedExceptionAsync(response).ConfigureAwait(false);
             }
         }
-        private static bool CheckCompleted(Response response)
+        private bool CheckCompleted(Response response)
         {
             switch (response.Status)
             {
@@ -132,7 +132,7 @@ namespace Azure.Security.KeyVault.Secrets
                     return false;
 
                 default:
-                    throw response.CreateRequestFailedException();
+                    throw _pipeline.Diagnostics.CreateRequestFailedException(response);
             }
         }
     }
